pyproject.toml
[project]
name = "jus-scraper"
version = "0.1.0"
description = ""
authors = [
    {name = "Natan Reis",email = "natangon10@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "selenium (>=4.34.0,<5.0.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "pandas (>=2.3.0,<3.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "webdriver-manager (>=4.0.2,<5.0.0)"
]

[tool.poetry]
packages = [{include = "jus_scraper", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
